use int for max content size
git-svn-id: https://svn.code.sf.net/p/waterken/code/server/trunk@245 ec340d7e-3b38-0410-a49c-f1b72a8843da

+++ waterken/web_send/src/org/web_send/Entity.java
@@ -19,7 +19,7 @@ Entity implements Promise<Entity>, Powerless, Selfless, Serializable {
     /**
      * maximum number of bytes in {@link #content}
      */
-    static public final long maxContentSize = 64 * 1024;
+    static public final int maxContentSize = 64 * 1024;
 
     /**
      * MIME Media Type
